+++ packages/atomic/src/components/commerce/atomic-product-description/atomic-product-description.ts
@@ -66,7 +66,7 @@ export class AtomicProductDescription
     converter: booleanConverter,
     attribute: 'is-collapsible',
   })
-  public isCollapsible = true;
+  public isCollapsible = false;
 
   initialize() {
     this.validateProps();