+++ packages/atomic/src/components/search/atomic-search-interface/atomic-search-interface.tsx
@@ -42,7 +42,7 @@ import {
   errorSelector,
   firstSearchExecutedSelector,
   noResultsSelector,
-} from '../atomic-layout/search-layout';
+} from '../atomic-search-layout/search-layout';
 import {getAnalyticsConfig} from './analytics-config';
 import type {Bindings as _Bindings} from './interfaces';
 import {createSearchStore, SearchStore} from './store';
@@ -58,7 +58,6 @@ export type Bindings = _Bindings;
   tag: 'atomic-search-interface',
   styleUrl: 'atomic-search-interface.pcss',
   shadow: true,
-  assetsDirs: ['lang'],
 })
 export class AtomicSearchInterface
   implements StencilBaseAtomicInterface<SearchEngine>