+++ packages/samples/angular/src/app/app.module.ts
@@ -5,6 +5,7 @@ import {AppComponent} from './app.component';
 import {AppRoutingModule} from './app-routing.module';
 import {AtomicAngularPageComponent} from './atomic-angular-page/atomic-angular-page.component';
 import {AtomicAngularRecommendationPageComponent} from './atomic-angular-recommendation-page/atomic-angular-recommendation-page.component';
+import {AtomicAngularRGAPageComponent} from './atomic-angular-rga-page/atomic-angular-rga-page.component';
 import {AtomicAngularSearchPageComponent} from './atomic-angular-search-page/atomic-angular-search-page.component';
 import {AtomicCommerceAngularPageComponent} from './atomic-commerce-angular-page/atomic-commerce-angular-page.component';
 import {FieldLabelComponent} from './field-label/field-label.component';
@@ -15,6 +16,7 @@ import {LabelAndFieldValueComponent} from './label-and-field-value/label-and-fie
   declarations: [
     AppComponent,
     AtomicAngularPageComponent,
+    AtomicAngularRGAPageComponent,
     AtomicAngularSearchPageComponent,
     AtomicAngularRecommendationPageComponent,
     AtomicCommerceAngularPageComponent,