+++ packages/create-atomic/CHANGELOG.md
@@ -1,3 +1,9 @@
+## <small>1.41.1 (2025-08-27)</small>
+
+* fix(cli): fix plop choice type to simple list (#5968) ([b0a1ace](https://github.com/coveo/ui-kit/commits/b0a1ace)), closes [#5968](https://github.com/coveo/ui-kit/issues/5968)
+
+
+
 ## 1.41.0 (2025-08-07)
 
 * feat: Migrate monorepo build system from NX to Turborepo (#5733) ([e7e1237](https://github.com/coveo/ui-kit/commits/e7e1237)), closes [#5733](https://github.com/coveo/ui-kit/issues/5733)